python/pyproject.toml
[project]
name = "watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.38.0",
    "mss>=10.1.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyaudio>=0.2.14",
]
